Use 32bits bitmapsets
This makes it possible to reduce the number of bits also in other
variables as well since, with up to 32 relations, fewer combinations
are possible.

+++ src/backend/optimizer/gpuqo/gpuqo.cuh
@@ -58,12 +58,12 @@ struct JoinRelation{
 	RelationID id;
 	RelationID left_relation_id;
 	union{
-		uint64_t left_relation_idx;
+		uint32_t left_relation_idx;
 		struct JoinRelation* left_relation_ptr;
 	};
 	RelationID right_relation_id;
 	union{
-		uint64_t right_relation_idx;
+		uint32_t right_relation_idx;
 		struct JoinRelation* right_relation_ptr;
 	};
 	double rows;
